+++ lib/World.lua
@@ -474,7 +474,7 @@ end
 	Only one changed record is returned per entity, even if the same entity changed multiple times. The order
 	in which changed records are returned is not guaranteed to be the order that the changes occurred in.
 
-	It should be noted that `queryChanged` does not have the same iterator invalidation limitations as `World:query`.
+	It should be noted that `queryChanged` does not have the same iterator invalidation concerns as `World:query`.
 
 	:::caution
 	The first time your system runs (i.e., on the first frame), no results are returned. Results only begin to be
@@ -490,33 +490,22 @@ end
 	every frame, **until the end of time**.
 	:::
 
-	### Arguments
-
-	The first argument to `queryChanged` is the component for which you want to track changes.
-	Further arguments are optional, and if passed, are an additional filter on what entities will be returned.
-
-	:::caution
-	Additional query arguments are checked against *at the time of iteration*, not when the change ocurred.
-	This has the additional implication that entities that have been despawned will never be returned from
-	`queryChanged` if additional query arguments are passed, because the entity will have no components, so cannot
-	possibly pass any additional query.
-	:::
-
-	If no additional query arguments are passed, all changes (including despawns) will be tracked and returned.
-
 	### Returns
 	`queryChanged` returns an iterator function, so you call it in a for loop just like `World:query`.
 
-	The iterator returns the entity ID, followed by a [`ChangeRecord`](#ChangeRecord), followed by the component
-	instance values of any additional query arguments that were passed (as discussed above).
+	The iterator returns the entity ID, followed by a [`ChangeRecord`](#ChangeRecord).
 
-	The ChangeRecord type is a table that contains two fields, `new` and `old`, respectively containing the new
+	The `ChangeRecord` type is a table that contains two fields, `new` and `old`, respectively containing the new
 	component instance, and the old component instance. `new` and `old` will never be the same value.
 
 	`new` will be nil if the component was removed (or the entity was despawned), and `old` will be nil if the
 	component was just added.
 
-	The ChangeRecord table is given to all systems tracking changes for this component, and cannot be modified.
+	The `old` field will be the value of the component the last time this system observed it, not
+	necessarily the value it changed from most recently.
+
+	The `ChangeRecord` table is potentially shared with multiple systems tracking changes for this component, so it
+	cannot be modified.
 
 	```lua
 	for id, modelRecord, enemy in world:queryChanged(Model, Enemy) do
@@ -539,10 +528,13 @@ end
 	:::
 
 	@param componentToTrack Component -- The component you want to listen to changes for.
-	@param ...? Component -- Additional query components. Checked at time of iteration, not time of change.
 	@return () -> (id, ChangeRecord, ...ComponentInstance) -- Iterator of entity ID followed by the requested component values, in order
 ]=]
 function World:queryChanged(componentToTrack, ...)
+	if ... then
+		error("World:queryChanged does not take any additional parameters", 2)
+	end
+
 	local hookState = TopoRuntime.useHookState(componentToTrack)
 
 	if not hookState.storage then
@@ -556,47 +548,15 @@ function World:queryChanged(componentToTrack, ...)
 		table.insert(self._changedStorage[componentToTrack], storage)
 	end
 
-	local queryLength = select("#", ...)
-	local queryOutput = table.create(queryLength)
-	local queryMetatables = { ... }
-
-	if #queryMetatables == 0 then
-		return function()
-			local entityId, component = next(hookState.storage)
-
-			if entityId then
-				hookState.storage[entityId] = nil
-
-				return entityId, component
-			end
-		end
-	end
-
-	local function queryIterator()
+	return function()
 		local entityId, component = next(hookState.storage)
 
 		if entityId then
 			hookState.storage[entityId] = nil
 
-			-- If the entity doesn't currently contain the requested components, don't return anything
-			if not self:contains(entityId) then
-				return queryIterator()
-			end
-
-			for i, queryMetatable in ipairs(queryMetatables) do
-				local queryComponent = self:get(entityId, queryMetatable)
-				if not queryComponent then
-					return queryIterator()
-				end
-
-				queryOutput[i] = queryComponent
-			end
-
-			return entityId, component, unpack(queryOutput, 1, queryLength)
+			return entityId, component
 		end
 	end
-
-	return queryIterator
 end
 
 function World:_trackChanged(metatable, id, old, new)